光が当たっている範囲の境界にエッジ

+++ HANAHIME/Game/resource/user/shaders/BasicShader_Huge.hlsl
@@ -133,7 +133,7 @@ struct PSOutput
     float depth : SV_Target2;
     float4 normal : SV_Target3;
     float4 edgeColor : SV_Target4;
-    uint4 grass : SV_Target5;
+    float4 grass : SV_Target5;
 };
 
 PSOutput PSmain(VSOutput input) : SV_TARGET
